+++ src/Facades/FieldType.php
@@ -5,10 +5,10 @@ namespace Ycs77\LaravelFormFieldType\Facades;
 use Illuminate\Support\Facades\Facade;
 
 /**
- * @method static array type(string $type, array $data = [])
- * @method static array field(array|string $value)
+ * @method static array type(string $name, array $data = [])
  * @method static array fields(array $fields)
  * @method static array list(array $fields)
+ * @method static array map(array $fields, Closure $collback)
  * @method static array casts(array $fields, array $data)
  * @method static \Kris\LaravelFormBuilder\Form render(\Kris\LaravelFormBuilder\Form $form, array $fields_data)
  * @method static array rules(array $fields)